some initial clean up
Initial clean up for the existing socket server to make it more generic and moving it out of one project to the other. Some boilerplate code as well for the initial core dump debugging but nothing meaningful at this point.
Also verified that existing functionality around the terminal is working fine for now.