converted to wizard
Converted the existing page to wizard and added the required steps
Git & Python configurable through wizard.
Download or browse to existing IDF tools path.

The manager for the tools installation needs some changes coming up in next commit.

This is not a complete commit of the feature but the code is mostly clean here