IEP-901: Adding descriptive port name from the jSerialComm API
The only name we can get from the api is descriptive port name which may not correspond to the correct device name but is good enough.
I have also added a separate thread to update the serial ports info so when a new device gets connected we simply update the serial ports list and also show a message in the information area for that.