IEP-965: Bug fix for serial port exception (#772)
The serial ports descriptor was initially split by the hyphen(-) which can be a valid serial port name char as well so instead now we are using the good old trust space( ) char to keep it clean and simple. Also did some ui improvements to help keep the user interface active while background tasks can take a while to resolve.